FEAT: Jump and attributes tweakes
The Jump mechanic was added, I changed WASD omni-directional movement
for just horizontal movement with the jump and have been trying the
horizontal speed, the force of jump and the gravity scale on player
object trying to get the best sensation for the game. The results were
decent and some platforms were added along with a ground to test the
movement and interaction with enviroment. And finally the rotation
property of the rigidbody of player object was constrained in order to
avoid rotation on collission.