FEAT: Jump Input Buffer
The jump felt strange and required a lot of accuracy and timing to get
the jump right, so it obviously needed an input buffer for the jump. The
jump script was reworked with the implementation of an input buffer, as
tested the jump now feels better but there is still problems with the
horizontal collissions.