REFACTOR/CHANGE: Unity Editor Version Downgrade
The computer I use on my education center only has installed the version
6000.0.50f1, while I had installed the version 6000.2 in my personal
computer, at the lack of admin permissions I had to downgrade the
project to match the version of the other computer, that caused several
troubles with refactoring, assets, render, packages, etc. So this change
is the stable version of this downgrade with everything seem up to
normal and without any compiler errors or warnings. Also minor changes
on the SampleScene: the player was moved so that it didn't collide with
a platform at the start of the scene and the ground platform had to be
wided a bit, as it seems that either the camera got smaller or the 2d
objects did.