FEAT: Shoot Bullet
I made the scripts for the bullet and the staff so that it shoots the
bullet whenever you left-click, also I made prefabs of most of the
important objects made. Everything seems to be working with the bullets
getting shot in the right direction and destroying themselves after the
indicated time, it might need to do some tweaks on the properties of the
bullet to improve the gameplay and feeling in general.