FEATURE: Coyote Time and tweaks on jump
I added coyote time so that the player is able to jump some miliseconds
after falling from the edge of a platform. Also I changed the force of
the jump to 20 and the gravity scale to 6, I think the jump feels better
as it is not much higher but it falls down faster.