ADD: Enemy
Created an enemy Prefab which is just a rectangle equal to the player
but in a red collor, also made them to destroy after getting hit by the
bullet thrice. For that I added the tag "Bullet" to the bullet prefab in
order for the enemy to detect its collision.